Planemodel finished
Planemodel has an id now same as plane.
Planemodel id i foreign key for plane.
planemodel id is primary key now.
Changed the ais.script (adjustments for plane and plane model).
Plane has seats now and the view includes this.
Planemodel needs to exist in database to insert the plane with the picked planemodel.
The planemodels in the addPlaneview are hardcoded.